Fix BCM mode test cases
The "rpi-gpio handles pin translation when in BCM mode writing to BCM pin X"
tests are currently passing although in actual use BCM mode fails on pins
where the BCM pin number isn't also a valid RPI pin number (issue #11).
The tests are passing because the test inputs are RPI pin numbers and not
BCM numbers and hence don't trigger the bug.

To give a concrete example: BCM pin 4 wasn't in the test input list although
it is of course a valid BCM pin. The same integer in the RPI pin space, RPI
pin 4, is not a valid pin.

This patch uses the BCM pin numbers instead of the RPI pin numbers. In
addition it repeats the tests for all hardware revisions therefore covering
both the v1 and v2 pin layouts which have different valid BCM pin numbers.

This patch also adds tests to check that using invalid BCM pin numbers
generates an error, again with cases for both the v1 and v2 layouts.