Fix the "rpi-gpio setup() when given a non-GPIO channel" test
This test should be run with a BCM pin number that is not valid. It was
being run with pin 1 and with a hardware revision that uses the v1 pin
layout. BCM pin 1 is a vlida pin in the v1 layout and is thus not suitable
for this test. The test has been passing because of issue #11 which
erroneously returns an error for some valid pins in BCM mode.

This patch resolves the test bug by using BCM pin 2 which is not a valid
pin in the v1 layout.